+++ gnucash/gnome-utils/gnc-tree-control-split-reg.c
@@ -36,9 +36,9 @@
 #include "gnc-tree-util-split-reg.h"
 #include "gnc-tree-view-split-reg.h"
 #include "gnc-component-manager.h"
+#include "gnc-date.h"
 #include "gnc-ui.h"
 #include "gnc-prefs.h"
-#include "gnc-gdate-utils.h"
 #include "gnc-warnings.h"
 #include "dialog-utils.h"
 #include "dialog-dup-trans.h"
