+++ gnucash/gnome/assistant-acct-period.c
@@ -45,9 +45,9 @@
 #include "assistant-acct-period.h"
 #include "gnc-component-manager.h"
 #include "qof.h"
+#include "gnc-date.h"
 #include "gnc-file.h"
 #include "gnc-frequency.h"
-#include "gnc-gdate-utils.h"
 #include "gnc-gui-query.h"
 #include "gnc-ui-util.h"
 #include "misc-gnome-utils.h"
