+++ gnucash/gnome/assistant-loan.c
@@ -41,7 +41,6 @@
 #include "dialog-utils.h"
 #include "Account.h"
 #include "gnc-ui.h"
-#include "gnc-gdate-utils.h"
 #include "gnc-gui-query.h"
 #include "gnc-ui-util.h"
 #include "gnc-frequency.h"
