+++ gnucash/gnome/dialog-invoice.c
@@ -32,13 +32,13 @@
 
 #include "qof.h"
 
-#include <gnc-gdate-utils.h>
 #include "dialog-utils.h"
 #include "gnc-component-manager.h"
 #include "gnc-ui.h"
 #include "gnc-gui-query.h"
 #include "gnc-prefs.h"
 #include "gnc-ui-util.h"
+#include "gnc-date.h"
 #include "gnc-date-edit.h"
 #include "gnc-amount-edit.h"
 #include "gnucash-sheet.h"
