+++ libgnucash/engine/Makefile.am
@@ -40,7 +40,6 @@ libgncmod_engine_la_SOURCES = \
   gnc-engine.c \
   gnc-event.c \
   gnc-features.c \
-  gnc-gdate-utils.c \
   gnc-hooks.c \
   gnc-int128.cpp \
   gnc-lot.c \
@@ -125,7 +124,6 @@ gncinclude_HEADERS = \
   gnc-engine.h \
   gnc-event.h \
   gnc-features.h \
-  gnc-gdate-utils.h \
   gnc-hooks.h \
   gnc-numeric.h \
   gnc-numeric.hpp \
